requirements.txt
# Django后端
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient>=2.2.0
django-cors-headers==4.3.0
django-simpleui==2025.6.24
python-decouple==3.8
cryptography>=43.0.0
requests==2.31.0
Pillow>=10.0.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Streamlit前端
streamlit==1.29.0
pandas==2.1.3
plotly==5.18.0
